requirements.txt
Flask
tensorflow
numpy
Pillow
psycopg2-binary
